arzonuching/requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn==0.30.6
aiohttp==3.10.5
python-dotenv==1.0.1
paytechuz==0.6.2